+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Improvements
 
 - Added support for row validation using XSD schema using `rowValidationXSDPath` option when reading XML files with a row tag using `rowTag` option.
+- Improved SQL generation for `session.table().sample()` to generate a flat sql statement.
 
 ### Snowpark Local Testing Updates
 
@@ -92,6 +93,7 @@
 - Set the default value of the `index` parameter to `False` for `DataFrame.to_view`, `Series.to_view`, `DataFrame.to_dynamic_table`, and `Series.to_dynamic_table`.
 - Added `iceberg_version` option to table creation functions.
 - Reduced query count for many operations, including `insert`, `repr`, and `groupby`, that previously issued a query to retrieve the input data's size.
+- Show more helpful error messages to Snowflake Notebook users when Modin or pandas version does not match our requirements.
 
 #### Bug Fixes